move profile follow button to shared
