+++ components/ProfilePage/ProfileHeader.tsx
@@ -19,65 +19,26 @@ import {
 import { EditProfileButton } from "./EditProfileButton"
 import { OrgContactInfo } from "./OrgContactInfo"
 import { Profile } from "../db"
-import { FollowButton } from "./FollowButton"
+import { FollowButton } from "components/shared/FollowButton"
 
 export const ProfileHeader = ({
   isUser,
   isOrg,
   isMobile,
-  uid,
+  profile,
   profileid,
-  profile
+  uid
 }: {
   isUser: boolean
   isOrg: boolean
   isMobile: boolean
-  uid?: string
-  profileid: string
   profile: Profile
+  profileid: string
+  uid?: string
 }) => {
   const orgImageSrc = profile.profileImage
     ? profile.profileImage
     : "/profile-org-icon.svg"
-  const topicName = `org-${profileid}`
-  const subscriptionRef = collection(
-    firestore,
-    `/users/${uid}/activeTopicSubscriptions/`
-  )
-  const [queryResult, setQueryResult] = useState("")
-
-  const orgQuery = async () => {
-    const q = query(
-      subscriptionRef,
-      where("topicName", "==", `org-${profileid}`)
-    )
-    const querySnapshot = await getDocs(q)
-    querySnapshot.forEach(doc => {
-      // doc.data() is never undefined for query doc snapshots
-      setQueryResult(doc.data().topicName)
-    })
-  }
-
-  useEffect(() => {
-    uid ? orgQuery() : null
-  })
-
-  const handleFollowClick = async () => {
-    await setDoc(doc(subscriptionRef, topicName), {
-      topicName: topicName,
-      uid: uid,
-      profileid: profileid,
-      type: "org"
-    })
-
-    setQueryResult(topicName)
-  }
-
-  const handleUnfollowClick = async () => {
-    await deleteDoc(doc(subscriptionRef, topicName))
-
-    setQueryResult("")
-  }
 
   return (
     <Header className={`gx-0 edit-profile-header`}>
@@ -100,19 +61,7 @@ export const ProfileHeader = ({
               {isUser ? (
                 <EditProfileButton isOrg={isOrg} isMobile={isMobile} />
               ) : (
-                /*
-                 remove comment when Notification Emails and related Follow functionality
-                 is ready for production
-                */
-
-                // <FollowButton
-                //   onFollowClick={() => handleFollowClick()}
-                //   onUnfollowClick={() => handleUnfollowClick()}
-                //   isMobile={isMobile}
-                //   isFollowing={queryResult}
-                // />
-
-                <></>
+                <FollowButton profileid={profileid} uid={uid} />
               )}
             </>
           )}
